requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
Flask==1.1.1
awscli==1.18.56
boto3==1.13.7
botocore==1.16.6
certifi==2020.4.5.1
docutils==0.15.2
jmespath==0.9.5
pyasn1==0.4.8
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3.1
PyMySQL==0.9.3
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.9
wincertstore==0.2
implicit==0.4.2
scikit-learn==0.21.2
numpy==1.18.1
pandas==1.0.3
requests==2.13.0
pytest==5.4.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.0

